+++ src/Sql/Sql/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Made `Start-AzSqlInstanceDatabaseLogReplay` cmdlet synchronous, added -AsJob flag
+
+## Version 2.14.0
 * Fixed parameter description for `InstanceFailoverGroup` command.
 * Updated the logic in which schemaName, tableName and columnName are being extracted from the id of SQL Data Classification commands.
 * Fixed Status and StatusMessage fields in `Get-AzSqlDatabaseImportExportStatus` to conform to documentation